+++ webknossos/webknossos/client/api_client/tracingstore_api_client.py
@@ -5,6 +5,8 @@ from webknossos.client.api_client.models import (
     ApiPrecomputedMeshInfo,
 )
 
+from ...proofreading.agglomerate_graph import AgglomerateGraph
+from ...proofreading.generated import agglomerate_graph_pb2
 from ._abstract_api_client import AbstractApiClient, Query
 
 
@@ -41,3 +43,13 @@ class TracingStoreApiClient(AbstractApiClient):
             body_structured=mesh,
             query=query,
         )
+
+    def get_agglomerate_graph(
+        self, tracing_id: str, agglomerate_id: int
+    ) -> AgglomerateGraph:
+        route = f"/mapping/{tracing_id}/agglomerateGraph/{agglomerate_id}"
+        agglomerate_graph_proto = self._get_parsed_protobuf(
+            route, agglomerate_graph_pb2.AgglomerateGraph
+        )
+        agglomerate_graph = AgglomerateGraph.from_proto(agglomerate_graph_proto)
+        return agglomerate_graph