+++ webknossos/webknossos/annotation/annotation.py
@@ -78,6 +78,7 @@ from ..dataset_properties import (
     VoxelSize,
 )
 from ..geometry import NDBoundingBox, Vec3Int
+from ..proofreading.agglomerate_graph import AgglomerateGraph
 from ..skeleton import Skeleton
 from ..utils import get_executor_for_args, is_fs_path, time_since_epoch_in_ms
 from ._nml_conversion import annotation_to_nml, nml_to_skeleton
@@ -981,7 +982,6 @@ class Annotation:
             dataset_id=self.dataset_id,
             annotation_id_or_url=self.annotation_id,
             use_zarr_streaming=True,
-            read_only=True,
         )
 
     def get_remote_base_dataset(
@@ -1473,6 +1473,7 @@ class RemoteAnnotation(Annotation):
                 modified=annotation_info.modified,
                 data_store=annotation_info.data_store,
                 tracing_time=annotation_info.tracing_time,
+                annotation_layers=None,
             ),
         )
 
@@ -1532,3 +1533,42 @@ class RemoteAnnotation(Annotation):
             for chunk in mesh_download:
                 f.write(chunk)
         return file_path
+
+    def get_agglomerate_graph(self, agglomerate_id: int) -> AgglomerateGraph:
+        """
+        Get the agglomerate graph for the specified agglomerate id.
+        This works only for proofreading annotations that have only a single volume layer.
+
+        Args:
+            agglomerate_id (int): The id of the agglomerate to get the graph for.
+
+        Returns:
+            AgglomerateGraph: The agglomerate graph for the specified agglomerate id.
+            The agglomerate graph has a vertex for all segments that belong to the agglomerate.
+            Adjacent segments are connected by an edge.
+
+        Raises:
+            ValueError: If the agglomerate id is not valid
+            UnexpectedStatusError: If the annotation does not have an editable mapping (is not a proofreading annotation)
+            AssertionError: If the annotation does not have exactly one volume layer
+
+        """
+        if agglomerate_id < 0:
+            raise ValueError("agglomerate_id must be a positive integer")
+
+        from ..client.context import _get_context
+
+        context = _get_context()
+        annotation_info = self._get_annotation_info()
+        assert annotation_info.annotation_layers is not None, "Annotation has no layers"
+        tracingstore_client = context.get_tracingstore_api_client()
+        volume_layer = [
+            layer
+            for layer in annotation_info.annotation_layers
+            if layer.typ == "Volume"
+        ]
+        assert len(volume_layer) == 1, "Expected exactly one volume layer"
+        graph = tracingstore_client.get_agglomerate_graph(
+            volume_layer[0].tracing_id, agglomerate_id
+        )
+        return graph